fb: notification transparency (1/2)
this add back notification transparency...compared to 4.3
it controlls as well the notification icon backgound transparency
which felt in the past dissplaced. As well codecleanup of the old
implementation.

largeIcon support I did not add due that google does not support it
as well at current status...it is only a placeholder in
NotificationData

Change-Id: Ia2c6761575dcf62adc5eae248125d4e0ad67be5e

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java